requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
autocorrect==2.5.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blis==0.7.4
brotlipy==0.7.0
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cloudpickle==1.6.0
colorama==0.4.4
contractions==0.0.49
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
en-core-web-sm==2.3.1
entrypoints==0.3
et-xmlfile==1.0.1
gitdb==4.0.7
GitPython==3.1.19
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.3.4
ipympl==0.7.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.17.2
Jinja2==3.0.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.0.2
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-resource-usage==0.6.0
jupyter-server==1.4.1
jupyter-server-mathjax==0.2.3
jupyterlab==3.0.16
jupyterlab-git==0.30.1
jupyterlab-kite==2.0.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-system-monitor==0.8.0
jupyterlab-topbar==0.6.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lckr-jupyterlab-variableinspector==3.0.9
lemminflect==0.2.2
librosa==0.8.0
llvmlite==0.36.0
MarkupSafe==2.0.0
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
mpmath==1.2.1
murmurhash==1.0.5
nb-conda-kernels==2.3.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbdime==3.1.0
nbformat==5.1.3
nbresuse==0.4.0
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numba==0.53.1
numpy==1.20.2
olefile==0.46
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
plac==1.1.0
pooch==1.3.0
preshed==3.0.5
prometheus-client==0.10.1
prompt-toolkit==3.0.17
psutil==5.8.0
ptyprocess==0.7.0
pyahocorasick==1.4.0
pycparser==2.20
Pygments==2.8.1
pympi-ling==1.69
PyOpenGL==3.1.1a1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypevoc==0.1
Pyphen==0.10.0
pyqtgraph==0.11.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.15.0
smmap==3.0.5
sniffio==1.2.0
SoundFile==0.10.3.post1
spacy==2.3.5
spyder-kernels==2.0.5
srsly==1.0.5
statsmodels==0.12.2
terminado==0.9.4
testpath==0.4.4
textsearch==0.0.21
tgt==1.4.4
thinc==7.4.5
threadpoolctl==2.1.0
timeseries==0.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==2.0.1
zipp==3.4.1
